+++ assembly-syntax/src/testing/context.rs
@@ -1,6 +1,6 @@
 use alloc::{boxed::Box, sync::Arc, vec::Vec};
 
-use miden_debug_types::{DefaultSourceManager, SourceFile, SourceManager};
+use miden_debug_types::{DefaultSourceManager, SourceFile, SourceManagerSync};
 use miden_utils_diagnostics::{
     Report,
     reporting::{ReportHandlerOpts, set_hook},
@@ -22,7 +22,7 @@ use crate::{
 /// Some of the assertion macros defined in this crate require a [SyntaxTestContext], so be aware of
 /// that.
 pub struct SyntaxTestContext {
-    source_manager: Arc<dyn SourceManager>,
+    source_manager: Arc<dyn SourceManagerSync>,
     warnings_as_errors: bool,
 }
 
@@ -66,7 +66,7 @@ impl SyntaxTestContext {
     }
 
     #[inline(always)]
-    pub fn source_manager(&self) -> Arc<dyn SourceManager> {
+    pub fn source_manager(&self) -> Arc<dyn SourceManagerSync> {
         self.source_manager.clone()
     }