+++ miden-vm/tests/integration/operations/decorators/mod.rs
@@ -1,6 +1,7 @@
 use std::sync::Arc;
 
 use miden_core::DebugOptions;
+use miden_debug_types::{DefaultSourceManager, Location, SourceFile, SourceManager, SourceSpan};
 use miden_processor::{
     AdviceMutation, AsyncHost, BaseHost, EventError, ExecutionError, MastForest, ProcessState,
     SyncHost,
@@ -18,9 +19,19 @@ pub struct TestHost {
     pub event_handler: Vec<u32>,
     pub trace_handler: Vec<u32>,
     pub debug_handler: Vec<String>,
+    pub source_manager: Arc<DefaultSourceManager>,
 }
 
 impl BaseHost for TestHost {
+    fn get_label_and_source_file(
+        &self,
+        location: &Location,
+    ) -> (SourceSpan, Option<Arc<SourceFile>>) {
+        let maybe_file = self.source_manager.get_by_uri(location.uri());
+        let span = self.source_manager.location_to_span(location.clone()).unwrap_or_default();
+        (span, maybe_file)
+    }
+
     fn on_debug(
         &mut self,
         _process: &mut ProcessState,