+++ processor/src/host/mod.rs
@@ -4,6 +4,7 @@ use core::future::Future;
 use miden_core::{
     AdviceMap, DebugOptions, Felt, Word, crypto::merkle::InnerNodeInfo, mast::MastForest,
 };
+use miden_debug_types::{Location, SourceFile, SourceSpan};
 
 use crate::{EventError, ExecutionError, ProcessState};
 
@@ -45,6 +46,12 @@ pub enum AdviceMutation {
 pub trait BaseHost {
     // REQUIRED METHODS
     // --------------------------------------------------------------------------------------------
+
+    /// Returns the [`SourceSpan`] and optional [`SourceFile`] for the provided location.
+    fn get_label_and_source_file(
+        &self,
+        location: &Location,
+    ) -> (SourceSpan, Option<Arc<SourceFile>>);
 
     /// Handles the debug request from the VM.
     fn on_debug(